create a hidden space at the start of a row
Since some photos don't have the title on the image, we loose context when using #anchors to go directly to content.
This commit uses the :before pseudo class to add a hidden space the size with the size of the top bar to add some offset to the scroll, when using the #anchor.